Rename framework_utils to framework_helpers and move to helpers module
