+++ dmscripts/insert_g8_framework_results.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-from dmscripts.framework_utils import has_supplier_submitted_services, set_framework_result
+
+from dmscripts.helpers.framework_helpers import has_supplier_submitted_services, set_framework_result
 
 FRAMEWORK_SLUG = 'g-cloud-8'
